Update docker-compose.yml and docs for Compose V2
See https://docs.docker.com/compose/releases/migrate/: Compose V2 was
released in 2020 and V1 stopped receiving updates in July 2023.

The main difference in usage is that the Compose spec doesn't use the
`version` top-level element and `compose` is now a subcommand of
`docker` instead of a standalone program called `docker-compose`.